refactor: CoreDataStack을 직접 생성
