+++ CloudNotes/CloudNotes/CoreDate/CoreDataStack.swift
@@ -1,0 +1,39 @@
+//
+//  CoreDataStack.swift
+//  CloudNotes
+//
+//  Created by 홍정아 on 2021/09/16.
+//
+
+import Foundation
+import CoreData
+
+class CoreDataStack {
+    private let modelName: String
+    
+    init(modelName: String) {
+        self.modelName = modelName
+    }
+    
+    lazy var persistentContainer: NSPersistentContainer = {
+        let container = NSPersistentContainer(name: modelName)
+        container.loadPersistentStores(completionHandler: { (_, error) in
+            if let error = error as NSError? {
+                fatalError("Unresolved error \(error), \(error.userInfo)")
+            }
+        })
+        return container
+    }()
+    
+    lazy var context: NSManagedObjectContext = persistentContainer.viewContext
+
+    func saveContext () {
+        guard context.hasChanges else { return }
+        do {
+            try context.save()
+        } catch {
+            let nserror = error as NSError
+            fatalError("Unresolved error \(nserror), \(nserror.userInfo)")
+        }
+    }
+}
