+++ CloudNotes/CloudNotes/Enumeration/CoreData.swift
@@ -1,0 +1,13 @@
+//
+//  CoreData.swift
+//  CloudNotes
+//
+//  Created by 홍정아 on 2021/09/16.
+//
+
+import Foundation
+
+enum CoreData {
+    static let noteEntity = "Note"
+    static let modelName = "CloudNotes"
+}
