+++ CloudNotes/Note+CoreDataProperties.swift
@@ -2,7 +2,7 @@
 //  Note+CoreDataProperties.swift
 //  
 //
-//  Created by 홍정아 on 2021/09/14.
+//  Created by 홍정아 on 2021/09/16.
 //
 //
 
@@ -15,8 +15,8 @@ extension Note {
         return NSFetchRequest<Note>(entityName: "Note")
     }
 
-    @NSManaged public var title: String
     @NSManaged public var body: String
+    @NSManaged public var title: String
     @NSManaged public var lastModified: Double
 
 }
